Updated both client and server file
I got the queries to run in asmt8server.py when the user chooses a query in Asmt8client.py. The Asmt8client.py file still has the server code, but it is just commented out, so that a copy of the query code is in both files for testing purposes.